requirements.txt
pandas
matplotlib
torch
torchvision
tqdm
scikit-learn
seaborn
fastapi
uvicorn
pillow
python-multipart